extract DummySubsystem into mod dummy.
