+++ node/service/src/overseer.rs
@@ -23,6 +23,8 @@ use polkadot_node_core_candidate_validation::Config as CandidateValidationConfig
 use polkadot_node_core_chain_selection::Config as ChainSelectionConfig;
 use polkadot_node_core_dispute_coordinator::Config as DisputeCoordinatorConfig;
 use polkadot_node_network_protocol::request_response::{v1 as request_v1, IncomingRequestReceiver};
+#[cfg(any(feature = "malus", test))]
+pub use polkadot_overseer::dummy::DummySubsystem;
 pub use polkadot_overseer::{
 	metrics::Metrics, MetricsTrait,
 	AllSubsystems, BlockInfo, HeadSupportsParachains, Overseer, OverseerBuilder, OverseerConnector,
